requirements.txt
anyio==4.0.0
asgiref==3.8.1
certifi==2023.7.22
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
Django==5.0.3
django-crispy-forms==2.1
djangorestframework==3.15.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
instaloader==4.10.3
pillow==10.3.0
pygame==2.5.2
python-telegram-bot==20.6
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.1.0
